+++ src/systems.js
@@ -1,0 +1,9 @@
+import { addSystem } from './modules/stores.js';
+
+import pf2e from "./systems/pf2e.js";
+addSystem(pf2e);
+import dnd5e from "./systems/dnd5e.js";
+addSystem(dnd5e);
+import common from "./systems/common.js";
+addSystem(common);
+