+++ src/view/Notifications.svelte
@@ -3,9 +3,6 @@
 <script>
    export let elementRoot;
 
-   import { setting } from "../modules/settings.js";
-   import { moduleId, SETTINGS } from "../modules/constants.js";
-
    import { onMount } from "svelte";
    import toast, { Toaster } from "svelte-french-toast";