+++ CHANGELOG.md
@@ -3,10 +3,15 @@
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
 ## [Unreleased]
+## [0.4.14 - 0.5.0](https://github.com/averrin/alpha-suit/compare/0.4.14...0.5.0)
+Added:
+  * New Grid feature. Available for DM and players (you can disable it in the settings)
+  * [Premium] You can have more than one grid.
+
 ## [0.4.13 - 0.4.14](https://github.com/averrin/alpha-suit/compare/0.4.13...0.4.14)
 Added:
   * "Folders white list" option for the File Manager. Only these folders will be indexed.
-  * "On demand" mode for file indexing. Starts the process only on a window opening.
+  * "On demand" mode for file indexing. Starts the process only on a window opening. **Now this option is default**
 
 ## [0.4.12 - 0.4.13](https://github.com/averrin/alpha-suit/compare/0.4.12...0.4.13)
 Fixed: